colorRefの場所を移動
